+++ RELEASES.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 vNext
 -----
 
+### aoms
+
+- Added: Function `ConvertUnixMillisToTime()`
+
 ### aomc
 
 - Added: Struct `RawAttribute`